fix: upgrade all deps, node 12 support

+++ deps/libgeos/geos/include/geos/index/sweepline/SweepLineEvent.h
@@ -7,7 +7,7 @@
  *
  * This is free software; you can redistribute and/or modify it under
  * the terms of the GNU Lesser General Public Licence as published
- * by the Free Software Foundation. 
+ * by the Free Software Foundation.
  * See the COPYING file for more information.
  *
  **********************************************************************/
@@ -15,15 +15,16 @@
 #ifndef GEOS_INDEX_SWEEPLINE_SWEEPLINEEVENT_H
 #define GEOS_INDEX_SWEEPLINE_SWEEPLINEEVENT_H
 
+#include <cstddef>
 #include <geos/export.h>
 
 // Forward declarations
 namespace geos {
-	namespace index { 
-		namespace sweepline { 
-			class SweepLineInterval;
-		}
-	}
+namespace index {
+namespace sweepline {
+class SweepLineInterval;
+}
+}
 }
 
 namespace geos {
@@ -34,48 +35,48 @@ class GEOS_DLL SweepLineEvent {
 
 public:
 
-	enum {
-		INSERT_EVENT = 1,
-		DELETE_EVENT
-	};
+    enum {
+        INSERT_EVENT = 1,
+        DELETE_EVENT
+    };
 
-	SweepLineEvent(double x, SweepLineEvent *newInsertEvent,
-			SweepLineInterval *newSweepInt);
+    SweepLineEvent(double x, SweepLineEvent* newInsertEvent,
+                   SweepLineInterval* newSweepInt);
 
-	bool isInsert();
+    bool isInsert();
 
-	bool isDelete();
+    bool isDelete();
 
-	SweepLineEvent* getInsertEvent();
+    SweepLineEvent* getInsertEvent();
 
-	int getDeleteEventIndex();
+    size_t getDeleteEventIndex();
 
-	void setDeleteEventIndex(int newDeleteEventIndex);
+    void setDeleteEventIndex(size_t newDeleteEventIndex);
 
-	SweepLineInterval* getInterval();
+    SweepLineInterval* getInterval();
 
-	/**
-	 * ProjectionEvents are ordered first by their x-value, and then by their eventType.
-	 * It is important that Insert events are sorted before Delete events, so that
-	 * items whose Insert and Delete events occur at the same x-value will be
-	 * correctly handled.
-	 */
-	int compareTo(const SweepLineEvent *pe) const;
+    /**
+     * ProjectionEvents are ordered first by their x-value, and then by their eventType.
+     * It is important that Insert events are sorted before Delete events, so that
+     * items whose Insert and Delete events occur at the same x-value will be
+     * correctly handled.
+     */
+    int compareTo(const SweepLineEvent* pe) const;
 
-	//int compareTo(void *o) const;
+    //int compareTo(void *o) const;
 
 private:
 
-	double xValue;
+    double xValue;
 
-	int eventType;
+    int eventType;
 
-	/// null if this is an INSERT_EVENT event
-	SweepLineEvent *insertEvent;
+    /// null if this is an INSERT_EVENT event
+    SweepLineEvent* insertEvent;
 
-	int deleteEventIndex;
+    size_t deleteEventIndex;
 
-	SweepLineInterval *sweepInt;
+    SweepLineInterval* sweepInt;
 
 };
 
@@ -83,7 +84,7 @@ private:
 //typedef SweepLineEvent indexSweepLineEvent;
 
 struct GEOS_DLL  SweepLineEventLessThen {
-	bool operator() (const SweepLineEvent* first, const SweepLineEvent* second) const;
+    bool operator()(const SweepLineEvent* first, const SweepLineEvent* second) const;
 };
 
 //bool isleLessThen(SweepLineEvent *first, SweepLineEvent *second);
