+++ deps/libgeos/geos/include/geos/noding/SegmentIntersector.h
@@ -7,7 +7,7 @@
  *
  * This is free software; you can redistribute and/or modify it under
  * the terms of the GNU Lesser General Public Licence as published
- * by the Free Software Foundation. 
+ * by the Free Software Foundation.
  * See the COPYING file for more information.
  *
  **********************************************************************/
@@ -15,15 +15,16 @@
 #ifndef GEOS_NODING_SEGMENTINTERSECTOR_H
 #define GEOS_NODING_SEGMENTINTERSECTOR_H
 
+#include <cstddef>
 #include <geos/export.h>
 
 #include <geos/inline.h>
 
 // Forward declarations
 namespace geos {
-	namespace noding {
-		class SegmentString;
-	}
+namespace noding {
+class SegmentString;
+}
 }
 
 namespace geos {
@@ -48,36 +49,39 @@ class GEOS_DLL SegmentIntersector {
 
 public:
 
-	/**
-	 * This method is called by clients
-	 * of the SegmentIntersector interface to process
-	 * intersections for two segments of the SegmentStrings
-	 * being intersected.
-	 */
-	virtual void processIntersections(
-		SegmentString* e0,  int segIndex0,
-		SegmentString* e1,  int segIndex1)=0;
+    /**
+     * This method is called by clients
+     * of the SegmentIntersector interface to process
+     * intersections for two segments of the SegmentStrings
+     * being intersected.
+     */
+    virtual void processIntersections(
+        SegmentString* e0,  size_t segIndex0,
+        SegmentString* e1,  size_t segIndex1) = 0;
 
-	/**
-	 * \brief
-	 * Reports whether the client of this class
-	 * needs to continue testing all intersections in an arrangement.
-	 * 
-	 * @return true if there is not need to continue testing segments
-	 *
-	 * The default implementation always return false (process all intersections).
-	 */
-	virtual bool isDone() const {
-		return false;
-	}
+    /**
+     * \brief
+     * Reports whether the client of this class
+     * needs to continue testing all intersections in an arrangement.
+     *
+     * @return true if there is not need to continue testing segments
+     *
+     * The default implementation always return false (process all intersections).
+     */
+    virtual bool
+    isDone() const
+    {
+        return false;
+    }
 
-	virtual ~SegmentIntersector() 
-	{ }
+    virtual
+    ~SegmentIntersector()
+    { }
 
 protected:
 
-	SegmentIntersector() {}
- 
+    SegmentIntersector() {}
+
 };
 
 /// Temporary typedef for namespace transition
