+++ deps/libgeos/geos/include/geos/simplify/TaggedLineSegment.h
@@ -7,7 +7,7 @@
  *
  * This is free software; you can redistribute and/or modify it under
  * the terms of the GNU Lesser General Licence as published
- * by the Free Software Foundation. 
+ * by the Free Software Foundation.
  * See the COPYING file for more information.
  *
  **********************************************************************
@@ -19,9 +19,9 @@
  * NOTES: Use of this class by DP simplification algorithms
  * makes it useless for a TaggedLineSegment to store copies
  * of coordinates. Using pointers would be good enough here.
- * We don't do it to avoid having to break inheritance from 
+ * We don't do it to avoid having to break inheritance from
  * LineSegment, which has copies intead. Wheter LineSegment
- * itself should be refactored can be discussed. 
+ * itself should be refactored can be discussed.
  *  --strk 2006-04-12
  *
  **********************************************************************/
@@ -35,10 +35,10 @@
 
 // Forward declarations
 namespace geos {
-	namespace geom {
-		class Coordinate;
-		class Geometry;
-	}
+namespace geom {
+class Coordinate;
+class Geometry;
+}
 }
 
 namespace geos {
@@ -51,30 +51,29 @@ namespace simplify { // geos::simplify
  * Used to index the segments in a geometry and recover the segment locations
  * from the index.
  */
-class GEOS_DLL TaggedLineSegment: public geom::LineSegment
-{
+class GEOS_DLL TaggedLineSegment: public geom::LineSegment {
 
 public:
 
-	TaggedLineSegment(const geom::Coordinate& p0,
-			const geom::Coordinate& p1,
-			const geom::Geometry* parent,
-			unsigned int index);
+    TaggedLineSegment(const geom::Coordinate& p0,
+                      const geom::Coordinate& p1,
+                      const geom::Geometry* parent,
+                      size_t index);
 
-	TaggedLineSegment(const geom::Coordinate& p0,
-			const geom::Coordinate& p1);
+    TaggedLineSegment(const geom::Coordinate& p0,
+                      const geom::Coordinate& p1);
 
-	TaggedLineSegment(const TaggedLineSegment& ls);
+    TaggedLineSegment(const TaggedLineSegment& ls);
 
-	const geom::Geometry* getParent() const;
+    const geom::Geometry* getParent() const;
 
-	unsigned int getIndex() const;
+    size_t getIndex() const;
 
 private:
 
-	const geom::Geometry* parent;
+    const geom::Geometry* parent;
 
-	unsigned int index;
+    size_t index;
 
 };
