+++ deps/libgeos/geos/src/index/chain/MonotoneChainSelectAction.cpp
@@ -8,7 +8,7 @@
  *
  * This is free software; you can redistribute and/or modify it under
  * the terms of the GNU Lesser General Public Licence as published
- * by the Free Software Foundation. 
+ * by the Free Software Foundation.
  * See the COPYING file for more information.
  *
  **********************************************************************
@@ -28,11 +28,11 @@ namespace index { // geos.index
 namespace chain { // geos.index.chain
 
 void
-MonotoneChainSelectAction::select(MonotoneChain& mc, unsigned int start)
+MonotoneChainSelectAction::select(MonotoneChain& mc, size_t start)
 {
-	mc.getLineSegment(start, selectedSegment);
+    mc.getLineSegment(start, selectedSegment);
 
-	select(selectedSegment);
+    select(selectedSegment);
 }
 
 } // namespace geos.index.chain
